style(utils): rename perf-logger.ts -> perf-log.ts to match its perfLog export
