+++ lib/chrome/prelaunch.ts
@@ -3,7 +3,7 @@ import { findChrome } from './find-chrome.ts';
 import { spawnChrome } from './spawn-chrome.ts';
 import { killProcessGroup } from '../utils/kill-process-group.ts';
 import { CHROMIUM_ARGS } from './chromium-args.ts';
-import { perfLog } from '../utils/perf-logger.ts';
+import { perfLog } from '../utils/perf-log.ts';
 import * as Paths from '../commands/daemon/paths.ts';
 import type { ChromeHandle } from '../types.ts';
