+++ lib/setup/browser.ts
@@ -3,7 +3,7 @@ import { bindServerToPort } from './bind-server-to-port.ts';
 import { findChrome } from '../chrome/find-chrome.ts';
 import { CHROMIUM_ARGS } from '../chrome/chromium-args.ts';
 import { prelaunchPromise, shutdownPrelaunch } from '../chrome/prelaunch.ts';
-import { perfLog } from '../utils/perf-logger.ts';
+import { perfLog } from '../utils/perf-log.ts';
 import * as RunState from './run-state.ts';
 import type { Browser } from 'playwright-core';
 import type { HTTPServer } from '../servers/web.ts';
