+++ lib/utils/perf-log.ts
